libvpd: Remove build warnings
The following build warnings were removed.

In file included from src/dataitem.cpp:29:0:
./src/libvpd-2/dataitem.hpp: In copy constructor 'lsvpd::DataItem::DataItem(const lsvpd::DataItem&)':
./src/libvpd-2/dataitem.hpp:77:8: warning: 'lsvpd::DataItem::prefLevelUsed' will be initialized after [-Wreorder]
    int prefLevelUsed; /**< PreferenceLevel of source used to fill
        ^
./src/libvpd-2/dataitem.hpp:73:8: warning:   'int lsvpd::DataItem::packedLength' [-Wreorder]
    int packedLength;
        ^
src/dataitem.cpp:39:2: warning:   when initialized here [-Wreorder]
  DataItem::DataItem( const DataItem& copyMe ) : humanName( copyMe.humanName ),
  ^
In file included from src/dataitem.cpp:29:0:
./src/libvpd-2/dataitem.hpp: In constructor 'lsvpd::DataItem::DataItem()':
./src/libvpd-2/dataitem.hpp:73:8: warning: 'lsvpd::DataItem::packedLength' will be initialized after [-Wreorder]
    int packedLength;
        ^
./src/libvpd-2/dataitem.hpp:71:11: warning:   'std::string lsvpd::DataItem::dataValue' [-Wreorder]
    string dataValue;  /**< Actual value of this data item,
           ^
src/dataitem.cpp:44:2: warning:   when initialized here [-Wreorder]
  DataItem::DataItem( ) : packedLength( 0 ), dataValue( "" )
  ^
src/dataitem.cpp: In member function 'int lsvpd::DataItem::setValue(const string&, int, const char*, int)':
src/dataitem.cpp:144:49: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    while( ( pos = val.find( '\0' ) ) != string::npos ||
                                                 ^
src/dataitem.cpp:145:43: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     ( pos = val.find( '\n' ) ) != string::npos ||
                                           ^
src/dataitem.cpp:146:43: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     ( pos = val.find( '\r' ) ) != string::npos )
                                           ^
src/dataitem.cpp:152:32: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    for( i = 0; i < val.length( ) && isspace( val.at( i ) ); i++ )
                                ^
src/component.cpp: In member function 'bool lsvpd::Component::isChild(lsvpd::Component*)':
src/component.cpp:837:38: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < mChildren.size(); i++)
                                      ^
src/component.cpp: In member function 'bool lsvpd::Component::isChild(const string&)':
src/component.cpp:852:38: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < mChildren.size(); i++)
                                      ^
src/component.cpp: In member function 'bool lsvpd::Component::removeLeaf(int)':
src/component.cpp:1031:29: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   if( index > mLeaves.size( ) )
                             ^
src/helper_functions.cpp: In static member function 'static std::string lsvpd::HelperFunctions::parseString(const string&, int, std::string&)':
src/helper_functions.cpp:81:29: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   } while  ((pos != string::npos) && (count < str_pos));
                             ^
src/helper_functions.cpp: In static member function 'static bool lsvpd::HelperFunctions::matches(const string&, const string&)':
src/helper_functions.cpp:282:49: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    if ( string::npos == (end = s1.find("*", beg))) {
                                                 ^
src/helper_functions.cpp:284:61: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (string::npos != (z = s2.find(s1.substr(beg, end), 0))) {
                                                             ^
src/helper_functions.cpp: In static member function 'static void lsvpd::HelperFunctions::str2chr(char**, const string&)':
src/helper_functions.cpp:313:27: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    while (i < str2.length())
                           ^
src/system.cpp: In member function 'bool lsvpd::System::removeLeaf(int)':
src/system.cpp:611:29: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   if( index > mLeaves.size( ) )
                             ^
In file included from src/vpddbenv.cpp:24:0:
./src/libvpd-2/vpddbenv.hpp: In copy constructor 'lsvpd::VpdDbEnv::VpdDbEnv(const lsvpd::VpdDbEnv&)':
./src/libvpd-2/vpddbenv.hpp:58:11: warning: 'lsvpd::VpdDbEnv::mDbFileName' will be initialized after [-Wreorder]
    string mDbFileName;
           ^
./src/libvpd-2/vpddbenv.hpp:57:11: warning:   'std::string lsvpd::VpdDbEnv::mEnvDir' [-Wreorder]
    string mEnvDir;
           ^
src/vpddbenv.cpp:46:3: warning:   when initialized here [-Wreorder]
   VpdDbEnv::VpdDbEnv( const VpdDbEnv& copyMe ):
   ^
In file included from src/vpddbenv.cpp:24:0:
./src/libvpd-2/vpddbenv.hpp: In constructor 'lsvpd::VpdDbEnv::VpdDbEnv(const string&, const string&, bool)':
./src/libvpd-2/vpddbenv.hpp:58:11: warning: 'lsvpd::VpdDbEnv::mDbFileName' will be initialized after [-Wreorder]
    string mDbFileName;
           ^
./src/libvpd-2/vpddbenv.hpp:57:11: warning:   'std::string lsvpd::VpdDbEnv::mEnvDir' [-Wreorder]
    string mEnvDir;
           ^
src/vpddbenv.cpp:51:2: warning:   when initialized here [-Wreorder]
  VpdDbEnv::VpdDbEnv( const string& envDir, const string& dbFileName,

Acked-by: Suzuki K. Poulose <suzuki@in.ibm.com>

+++ src/component.cpp
@@ -834,7 +834,7 @@ lderr:
 	 */
 	bool Component::isChild(Component *child)
 	{
-		for (int i = 0; i < mChildren.size(); i++)
+		for (unsigned int i = 0; i < mChildren.size(); i++)
 			if (mChildren[i] == child->idNode.getValue())
 				return true;
 		return false;
@@ -849,7 +849,7 @@ lderr:
 	 */
 	bool Component::isChild(const string& child)
 	{
-		for (int i = 0; i < mChildren.size(); i++)
+		for (unsigned int i = 0; i < mChildren.size(); i++)
 			if (mChildren[i] == child)
 				return true;
 		return false;
@@ -1028,7 +1028,7 @@ lderr:
 	bool Component::removeLeaf( int index )
 	{
 		vector<Component*>::iterator i = mLeaves.begin( );
-		if( index > mLeaves.size( ) )
+		if( index > ( int )mLeaves.size( ) )
 		{
 			return false;
 		}
